requirements.txt
Flask~=3.1.2
python-dotenv~=1.1.1
pillow~=12.0.0
httpx~=0.28.0
openai~=1.0.0
flask-cors~=4.0.0

# Testing dependencies
pytest~=8.0.0
pytest-flask~=1.3.0
pytest-cov~=4.1.0
pytest-mock~=3.12.0
respx~=0.21.0
pytest-env~=1.1.0